Fix code review #JC19-355
