+++ apps/vanilla/src/pages/login/index.html
@@ -12,11 +12,11 @@
     <form class="form">
         <label class="form__label">
           <span class="form__label-text">Email: </span>
-          <input required name="email" type="email" placeholder="user@example.com" class="form__label-input form__label-email" />
+          <input required autocomplete="on" name="email" type="email" placeholder="user@example.com" class="form__label-input form__label-email" />
         </label>
         <label class="form__label">
           <span class="form__label-text">Password:</span>
-          <input required name="password" type="password" placeholder="Enter your password" class="form__label-input form__label-password" />
+          <input required autocomplete="on" name="password" type="password" placeholder="Enter your password" class="form__label-input form__label-password" />
         </label>
       <div class="form__submit">
         <button type="submit" class="form__submit-button">Sign in</button>
