+++ apps/vanilla/src/pages/register/index.html
@@ -13,7 +13,7 @@
       <div class="form__label">
         <label class="form__label">
           <span class="form__label-text">Email:</span>
-          <input name="email" type="email" required placeholder="user@example.com" class="form__label-input form__label-email" />
+          <input name="email" type="email" required autocomplete="on" placeholder="user@example.com" class="form__label-input form__label-email" />
         </label>
         <label class="form__label">
           <span class="form__label-text">First name:</span>
@@ -25,11 +25,11 @@
         </label>
         <label class="form__label">
           <span class="form__label-text">Password:</span>
-          <input name="password" required type="password" placeholder="Type your password" class="form__label-input form__label-password" />
+          <input name="password" required autocomplete="on" type="password" placeholder="Type your password" class="form__label-input form__label-password" />
         </label>
         <label class="form__label">
           <span class="form__label-text">Retype password:</span>
-          <input name="retype-password" required type="password" placeholder="Retype your password" class="form__label-input form__label-retypepassword" />
+          <input name="retype-password" required autocomplete="on" type="password" placeholder="Retype your password" class="form__label-input form__label-retypepassword" />
         </label>
       </div>
       <div class="form__submit">
